requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
plotly>=5.10.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.23.0
pillow>=9.0.0

